[PATCH] Initial backport of susfs v1.5.12 for kernel-5.4
This is an initial patch backporting susfs v1.5.12 (extracted from the gki-android12-5.10 branch) for kernel 5.4. It is specifically designed for google kernel 5.4; if it doesn't work for you, you will need to apply the patches manually. It does not require susfs v1.5.5, and it may fail if you already have it.